requirements.txt
fastapi
uvicorn
scikit-learn
keras
tensorflow
gunicorn
pillow
pymongo
keras